\name{Country Population}
\alias{Country Population}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Population growth over time in a specific country
}
\description{
Shows a country population over time for the year
}
\usage{
CountryPopulation(x)
}

\arguments{
  \item{x}{
A name of country
}
}

\author{
Louis B
}
